ignore_first_episode workaround for handling envs that can't be reset
procgen uses gym3, which doesn't allow early resetting of the
environment. This can mess with eval, which simply stops on getting
enough completed episodes irrespective of the done states of the envs.
This means that the next eval can start in the middle of an episode.
This plays out these first episodes ignoring their results.