Support playing procgen environments. Use IMPALA-style networks in Atari and CarRacing (#8)
* Initial procgen setup with installation commands

Does not successfully train with PPO

* Move FeatureExtractor to its own feature_extractor module

* fixup: Complete refactor move of FeatureExtractor

* ignore_first_episode workaround for handling envs that can't be reset

procgen uses gym3, which doesn't allow early resetting of the
environment. This can mess with eval, which simply stops on getting
enough completed episodes irrespective of the done states of the envs.
This means that the next eval can start in the middle of an episode.
This plays out these first episodes ignoring their results.

* IMPALA-style feature extractor

* Remove start_level argument

I was hoping it would act like setting an env seed, but it doesn't
appear to have made it deterministic (at least in eval)

* env_id hyperparam allows multiple items to point to the same env

This allows me to have separate specifications for easy and hard mode as
well as debug modes.

To more easily differentiate these meta items I've added an underscore
to the front of the names (_atari is also technically a meta name since
you can't specify _atari in the command line arguments, though how you
reach the actual env_id is different)

* Setup for Lambda Labs benchmarking session

* Add instruction to checkout branch before setup in lambda instructions

* Test out IMPALA-style network on ppo Breakout

* fixup impala-BreakoutNoFrameskip-v4 pulling hparams from base _atari

* Use the RunArg env name instead of env_id

The arg env name is assumed to have the necessary identifying information

* Refactor make_env code to use EnvHyperparams

This hopefully makes it easier to have branching make_env paths to
differentiate procgen and more normal gym env constructions

* procgen envs use ProcgenEnv and not stable-baselines3 VecEnv

This is a pretty major refactor that involved also making wrappers that
used to go around VecEnvs into just gym.Wrappers. This causes typing to
be much more difficult to enforce but generally seems to work.

* Try ImpalaCNN on NatureCNN environments. Add benchmarking scripts