Major refactor to use gym-like normalization instead of SB3's
The impetus came from huggingface complaining about the vecnormalize.pkl
file being unsafe. This seems to be an issue with the stable-baselines3
library, and it seemed awkward to rely on such a large library for
behavior that is already implemented in gym.

However, this wasn't entirely true given gym doesn't persist
normalization values, so I had to reimplement normalization. This
reimplemtation uses gym's SyncVectorEnv and AsyncVectorEnv instead of
sb3's VectorEnv environments.