Massive refactor moving src files into rl_algo_impls dir
This is to support packaging rl_algo_impls. When I attempted to
reinstall the dev environment using poetry there were a number of issues
with the older libraries (primarily from the use of an old version of
gym) not installing. That required me to switch to a requirements.txt,
which drove me to using a pyproject.toml file to use setuptools. The
fact this puts me on the road to packaging is a bonus.