score_threshold to determine best in eval
This is useful when self-play is the mechanism. For LuxAI_S2-v0, a score_threshold of 0.27 coincides with winning 64% of games, which given 66 games played is associated with a p>0.99. This isn’t actually correct given the opponent is a swath of past models, but it seems like a good approximation