Record results info & score based on cost+hp
The win_loss reward head was broken because microrts did not record info to results, this meant that runs would be broken as the reward would be pushed towards only using WinLoss (which was always 0 for microrts).

The score is the sum of the cost of the units with an additional term added for hp/max_hp. So the max score is twice the cost of the units if all units are at max_hp.

Additionally, a score_reward term is derived from the difference in scores divided by the sum of scores (plus a constant factor), which will give a reward range from -1 to 1. Microrts will have a third head that rewards based off of the change of this reward.