Microrts-finetuned-map16-from-NtR
Train a model for all the public maps up to size 16x16 (5 out of the 8 public maps) using the NoWhereToRun finetuned model. The NoWhereToRun model was trained from a general “up-to-16x16” model, which couldnt’ figure out NoWhereToRun. Given NoWhereToRun is almost 90% winning, try to transfer this knowledge back to the general case.