Use avg_ms from PGA for turn skipping insufficient time
Pass along the pre-game analysis expected average milliseconds to compute a time to the Java process. When RAISocketAI needs to wait for a step to compute on a following turn, it’s left with a decision to either compute the current turn or to skip the current turn because there isn’t enough time left. Use the pre-game analysis expected time to determine if there’s enough time (110% expected time or less is considered grounds to skip).