Revert "Use better performing model for map32"
This reverts commit 1bf93549bd7f70ee4018633003a833067904846a.

While the model performed better on DistantResources, it did not seem to
perform better on DoubleGame. Also the turning curve seemed much worse
than the old model implying the model might be adapted well for
DistantResources but not generally good. Additionally, DistantResources
now has a backup model based on squnet-map32, so having a fallback model
that only performs on DistantResources is no longer useful.