Fix RAISocketAI preventing script termination
ThreadPoolExecutor was holding onto two threads that prevented the object (and command-line program) from shutting itself down. The fix is to allow ThreadPoolExecutor to fall to 0 live threads once a timelimit is reached (5 seconds). If the executor is called again after this point, it’ll simply spin up a new thread.