ReferenceAIRollout is a way to accomplish offline learning
Required adding back in bot vs bot playing and then recording the action played by the bots.

This has a bug where log probabilities for actions don’t work because subactions for not applicable actions (think harvest direction during a move action) are selecting the 0-index, even if 0-index is invalid by the mask. The fix for this is to ignore not applicable subactions, which will likely help online version of PPO as well.