Actor-critic Behavior Cloning (#31)
This original branch was for a form of offline PPO; however, I couldn't get it to learn stably. The closest implemented version of what I wanted was BPPO (https://github.com/Dragon-Zhuang/BPPO), which did behavior cloning to bootstrap the policy and critic. ACBC (Actor-critic Behavior Cloning) is my implementation of behavior cloning that trains the actor and critic heads simultaneously.

ACBC was still unstable until I scaled the policy loss by the number of available positions to move. This scaling helped BPPO train without diverging, but OPPO is still unstable. Next steps is to train OPPO and PPO with the best ACBC-trained model so far.