Keep full batch off accelerator
Only transfer the minibatches to the accelerator (GPU or MPS). This is necessary when the map is 64x64 since the batch takes at least 22GB (15GB observation, 3GB actions, 4GB action mask). This could slow down smaller problems because the transfer to the accelerator can occur multiple times and occurs for each minibatch.