SummaryWrapper handles global_step counting for SummaryWriter
The idea is that the Algorithm is responsible for setting updating the global_steps count for everything.

I’m not certain if this is a good idea. It prevents EpisodeStatsWriter from diverging with the general step-count at least, but weirdly means it can record events at the same global_step.