Experiment: Try pushing rollout refs instead of rollouts
The hope is it keeps the data_store_actor from bloating in memory.

However, experimentally with APPO LunarLander-v2 locally it appears to
be about 20% slower. This is likely from the overhead of using the
object store and/or the overhead of getting the RemoteLearnerView and
then having to hit the object store to get the rollouts.