Microrts-env64-240m-ent5d1-lr8d1-05wb2lhr-2x2a100
Run DPPO in 2x2 GPU setup where each pair of GPUs is split between training and inference. This matches the way Lux map64 was trained; however, that was because the environment used Jax (Jux), which took a lot of GPU memory. This shouldn’t be necessary for Lux, but it might help against the crashes I’m observing when running 4x where each GPU is responsible for training and inference.